Refactor the SDK support for the MFA platform
Added two SDK classes - MPinSDK for the old SDK and MfaSDK for the new MFA flow.
Moved common functionality to a common base class MPinSDKBase.
Added customerId and appId to a User.
Changed users storage to be organized by mpinId only (and not by backend).

Closes: MAASMOB-436, MAASMOB-442
